Top-10 study: client-based crawl, plan fingerprints, timeline view, report fixes
- crawl lists through the authenticated client with spacing and backoff; ratings joined
  from the community index; refresh_ratings.py fills the rest slowly
- order-insensitive plan signature that ignores unexecutable market orders, plus an
  invalid-order count per game
- day-by-day timeline of the median game in every dossier
- branch-driver table at the first branching cut (weeds, opponent line, seat, shop draw)
- figure-safe team names, team-id slugs for non-Latin names
- journal and problems entries (throttling, RNG coupling, SpaTaro noise orders)

+++ research/sampling.py
@@ -18,7 +18,7 @@ def windows(n: int) -> dict[str, list[int]]:
     half = WINDOW // 2
     out = {"F": list(range(WINDOW))}
     for name, frac in (("Q1", 0.25), ("Q2", 0.5), ("Q3", 0.75)):
-        centre = int(round(n * frac))
+        centre = round(n * frac)
         start = min(max(centre - half, 0), n - WINDOW)
         out[name] = list(range(start, start + WINDOW))
     out["L"] = list(range(n - WINDOW, n))
